fix: :bug: Fix stdin stuck on redirection fd after builtin redirection
